packages: init packages directory; use `package.nix` as entry-point
This initializes a structure similar to the one found in nixpkgs (ie.
by-name). slippi and pplus have been purposefully left outside of them,
as the derivations under `by-name` can be used as-is.

The former being a base for other packages, and the latter not
functioning as of yet, cannot be evaluated the `callPackage` way.